+++ src/main/java/com/example/demo/Entities/NoticeSeen.java
@@ -15,11 +15,11 @@ public class NoticeSeen {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
-    private Integer noticeId;
+    private Long noticeId;
 
-    private Integer userId;
+    private Long userId;
 
     private String userRole; // NORMAL_USER / STAFF / etc
 
@@ -28,27 +28,27 @@ public class NoticeSeen {
 // GETTERS & SETTERS
 
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getNoticeId() {
+    public Long getNoticeId() {
         return noticeId;
     }
 
-    public void setNoticeId(Integer noticeId) {
+    public void setNoticeId(Long noticeId) {
         this.noticeId = noticeId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
