+++ src/main/java/com/example/demo/Payloads/CompanyDto.java
@@ -4,7 +4,7 @@ import com.example.demo.Enums.CompanyType;
 
 public class CompanyDto {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String logoUrl;
     private CompanyType companyType;
@@ -12,11 +12,12 @@ public class CompanyDto {
 
 // GETTERS & SETTERS
 
-    public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
