+++ src/main/java/com/example/demo/Payloads/NormalUserProfileDto.java
@@ -5,7 +5,7 @@ import com.example.demo.Enums.UserStatus;
 
 public class NormalUserProfileDto {
 
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
     private String mobileNumber;
@@ -14,8 +14,8 @@ public class NormalUserProfileDto {
 
     private UserStatus userStatus;
 
-    // getters & setters
 
+// GETTERS & SETTERS
 
     public UserStatus getUserStatus() {
         return userStatus;
@@ -25,11 +25,11 @@ public class NormalUserProfileDto {
         this.userStatus = userStatus;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
